feat(perps): implement withdraw (handle_unlock_liquidity)
Add the withdraw function that burns vault shares and creates an Unlock
record releasing settlement currency after a cooldown period. This is
the inverse of the deposit function.

- Move shared constants (VIRTUAL_SHARES, VIRTUAL_ASSETS, BANK, ORACLE)
  from deposit.rs to execute.rs for reuse across both modules
- Add checked_sub, checked_sub_assign, and is_non_zero to BaseAmount
- Change max_unlocks and max_open_orders from u32 to usize
- Include 13 unit tests covering symmetric withdrawal, partial, zero
  shares, unexpected coins, catastrophic loss, insufficient margin,
  max unlocks, unrealized PnL, funding, multiple pairs, non-dollar
  settlement price, end time, and floor rounding

+++ dango/types/src/perps.rs
@@ -58,11 +58,11 @@ pub struct Param {
     pub vault_cooldown_period: Duration,
 
     /// Maximum number of unlock requests a single user may have.
-    pub max_unlocks: u32,
+    pub max_unlocks: usize,
 
     /// Maximum number of resting limit order a single user may have across all
     /// trading pairs.
-    pub max_open_orders: u32,
+    pub max_open_orders: usize,
 
     /// Trading fee as a fraction of an order's notional value, deducted from
     /// the user's margin and transferred to the vault on every fill.
